Refactor SSE client and add unit tests
Extracted the SSE client logic into a reusable `mcp` package. Added unit tests to cover connection flow and retry mechanisms, improving maintainability and test coverage.
